backend/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
sqlalchemy==2.0.49
pydantic==2.13.3
pydantic-settings==2.14.0
python-multipart==0.0.27
python-dotenv==1.2.2
openai==1.10.0
httpx[socks]==0.28.1
aiofiles==25.1.0
PyJWT==2.12.1
bcrypt==5.0.0
alembic==1.18.4
sse-starlette==3.4.1

# Word Formatter dependencies
mistune>=3.0.0
python-docx>=1.1.0
lxml>=4.9.0
